Refactor following code review
- Standardise on using Newsletters in relevant class names.
- Move grouped and ungrouped newsletters agents into a single agent.
- Small formatting fixes.
