+++ applications/app/views/signup/newsletterContent.scala.html
@@ -1,7 +1,7 @@
 @import model.Page
 @import com.gu.identity.model.EmailNewsletter
 @import com.gu.identity.model.EmailNewsletters._
-@import services.newsletters.EmailSignupAgent
+@import services.newsletters.NewsletterSignupAgent
 @import services.newsletters.NewsletterResponse
 @import views.support.`package`.Seq2zipWithRowInfo
 
@@ -11,18 +11,20 @@
 @import services.newsletters.NewsletterResponse
 @(signupPage: Page)(implicit request: RequestHeader, context: model.ApplicationContext)
 
-@manageEmailsUrl = @{"https://manage.theguardian.com/email-prefs"}
+    @manageEmailsUrl = @{
+        "https://manage.theguardian.com/email-prefs"
+    }
 
-@emailListCategoryList(theme: String, emailListings: List[NewsletterResponse]) = {
-    <div class="newsletters-category__heading">
+    @emailListCategoryList(theme: String, emailListings: List[NewsletterResponse]) = {
+        <div class="newsletters-category__heading">
         @theme
-    </div>
-    <div class="newsletter-card__wrapper">
+        </div>
+        <div class="newsletter-card__wrapper">
         @emailListings.zipWithRowInfo.map { case (emailListing, row) =>
-            <div class="newsletter-card" data-component="newsletter-card @emailListing.id">
-                <div class="newsletter-card__content js-newsletter-content">
+        <div class="newsletter-card" data-component="newsletter-card @emailListing.id">
+            <div class="newsletter-card__content js-newsletter-content">
 
-                    <div class="newsletter-card__name">
+                <div class="newsletter-card__name">
                         @emailListing.name
                     </div>
 
@@ -78,29 +80,29 @@
                               </button>
                           </a>
                         }
-                    </div>
-                </div>
-            </div>
+        </div>
+        </div>
+        </div>
         }
-    </div>
-}
-
-@displayNewslettersByTheme = {
-    @signupPage.metadata.groupedNewslettersResponses.map { case (theme, newsletters) =>
-        <div class="newsletters-category" name="@theme">
-        @emailListCategoryList(theme, newsletters)
         </div>
     }
-}
 
-<div class="l-side-margins">
-<div class="content">
-    <div class="gs-container">
-        <div class="newsletters-heading__container">
-            <div class="newsletters-heading">
+    @displayNewslettersByTheme = {
+    @signupPage.metadata.groupedNewslettersResponses.map { case (theme, newsletters) =>
+    <div class="newsletters-category" name="@theme">
+    @emailListCategoryList(theme, newsletters)
+    </div>
+    }
+    }
+
+    <div class="l-side-margins">
+        <div class="content">
+            <div class="gs-container">
+                <div class="newsletters-heading__container">
+                    <div class="newsletters-heading">
                 Guardian newsletters: sign&nbsp;up
-            </div>
-            <a data-link-name="all-newsletters : manage-emails" href="@manageEmailsUrl">
+                    </div>
+                    <a data-link-name="all-newsletters : manage-emails" href="@manageEmailsUrl">
                 <div class="newsletters-heading__manage-button">
                     <span> Manage my newsletters </span>
                 </div>
