+++ common/app/services/newsletters/NewsletterApi.scala
@@ -41,6 +41,7 @@ object GroupedNewsletterResponse {
   implicit val groupedNewsletterResponseReads = Json.reads[GroupedNewsletterResponse]
 }
 
+// TODO: Find a better way to define this that means Frontend doesn't have knowledge of the fields returned.
 case class GroupedNewslettersResponse(
     newsRoundups: GroupedNewsletterResponse,
     newsByTopic: GroupedNewsletterResponse,
@@ -68,6 +69,19 @@ case class GroupedNewslettersResponse(
 
 object GroupedNewslettersResponse {
   implicit val groupedNewslettersResponseReads = Json.reads[GroupedNewslettersResponse]
+
+  // Create an empty response to initialise the box
+  val empty = GroupedNewslettersResponse(
+    GroupedNewsletterResponse("News roundups", Nil),
+    GroupedNewsletterResponse("News by topic", Nil),
+    GroupedNewsletterResponse("Features", Nil),
+    GroupedNewsletterResponse("Sport", Nil),
+    GroupedNewsletterResponse("Culture", Nil),
+    GroupedNewsletterResponse("Lifestyle", Nil),
+    GroupedNewsletterResponse("Comment", Nil),
+    GroupedNewsletterResponse("Work", Nil),
+    GroupedNewsletterResponse("From the papers", Nil),
+  )
 }
 
 case class NewsletterApi(wsClient: WSClient)(implicit executionContext: ExecutionContext)
