+++ dev-build/app/AppLoader.scala
@@ -31,13 +31,7 @@ import services._
 import _root_.commercial.targeting.TargetingLifecycle
 import akka.actor.ActorSystem
 import concurrent.BlockingOperations
-import services.newsletters.{
-  EmailEmbedLifecycle,
-  EmailSignupAgent,
-  GroupedNewslettersAgent,
-  GroupedNewslettersLifecycle,
-  NewsletterApi,
-}
+import services.newsletters.{NewsletterSignupLifecycle, NewsletterSignupAgent, NewsletterApi}
 import play.api.OptionalDevContext
 
 class AppLoader extends FrontendApplicationLoader {
@@ -60,8 +54,7 @@ trait Controllers
     with CricketControllers {
   self: BuiltInComponents =>
 
-  def emailEmbedAgent: EmailSignupAgent
-  def groupedNewslettersAgent: GroupedNewslettersAgent
+  def newsletterSignupAgent: NewsletterSignupAgent
 
   lazy val accessTokenGenerator = wire[AccessTokenGenerator]
   lazy val apiSandbox = wire[ApiSandbox]
@@ -88,8 +81,7 @@ trait AppComponents
   override lazy val contentApiClient = wire[ContentApiClient]
   override lazy val blockingOperations = wire[BlockingOperations]
   override lazy val newsletterApi = wire[NewsletterApi]
-  override lazy val emailEmbedAgent = wire[EmailSignupAgent]
-  override lazy val groupedNewslettersAgent = wire[GroupedNewslettersAgent]
+  override lazy val newsletterSignupAgent = wire[NewsletterSignupAgent]
 
   lazy val logbackOperationsPool = wire[LogbackOperationsPool]
 
@@ -122,8 +114,7 @@ trait AppComponents
       wire[TargetingLifecycle],
       wire[DiscussionExternalAssetsLifecycle],
       wire[StocksDataLifecycle],
-      wire[EmailEmbedLifecycle],
-      wire[GroupedNewslettersLifecycle],
+      wire[NewsletterSignupLifecycle],
     )
 
   override lazy val httpFilters = wire[DevFilters].filters
