+++ identity/app/AppLoader.scala
@@ -18,7 +18,7 @@ import play.api.mvc.EssentialFilter
 import play.api.routing.Router
 import play.filters.csrf.{CSRFAddToken, CSRFCheck, CSRFComponents}
 import router.Routes
-import services.newsletters.{EmailEmbedLifecycle, GroupedNewslettersLifecycle}
+import services.newsletters.NewsletterSignupLifecycle
 
 class AppLoader extends FrontendApplicationLoader {
   def buildComponents(context: Context): FrontendComponents =
@@ -42,8 +42,7 @@ trait AppLifecycleComponents {
     wire[IdentityLifecycle],
     wire[CloudWatchMetricsLifecycle],
     wire[SwitchboardLifecycle],
-    wire[EmailEmbedLifecycle],
-    wire[GroupedNewslettersLifecycle],
+    wire[NewsletterSignupLifecycle],
   )
 
   def actorSystem: ActorSystem
